+++ .gitignore
@@ -3,6 +3,9 @@
 # dependencies
 node_modules/
 
+#assets
+assets
+
 # Expo
 .expo/
 dist/
